refactor(components): move dashboard nav to main components folder
